requirements.txt
pycampbellcr1000
google-cloud-bigquery
rak811
python-dotenv
pyyaml
sqlite3
pyserial